Move new methods to ZooUtil
